refactor(ai): convert provider.rs to module directory
Closes #1350
